Create README.md
During the outbreak of covid19, everything in the world came to a standstill, that includes businesses and most especially schools. 
Later on, people got a way of maneuvering through it, but what happened to students? Most campus students
were greatly affected by it and didn't have a choice but to do what was asked of them.
The movement in and out of school greatly affected their studies. They had to study online and not all adhered to that, this system helps
the administration to get details about the whole institution. Including the lecturers and the students and the courses, they are taking.